feat: Simplify command syntax
At present, keywords (eg. todo, deadline) not only have to be written in
full but are also case sensitive.

This is not user friendly since case sensitivity is not meaningful and
the user should not have to type the entire command where there is no
ambiguity as to which command is desired.

Let's
* update the parser to be more flexible, by using case insensitive
  matching and allowing partial commands